backend/requirements/dev.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --cert=None --client-cert=None --index-url=None --output-file=dev.txt --pip-args=None dev.in
#
aiosqlite==0.20.0
    # via -r base.in
alembic==1.16.5
    # via -r base.in
annotated-types==0.7.0
    # via pydantic
anyio==4.10.0
    # via
    #   httpx
    #   openai
    #   starlette
asttokens==3.0.0
    # via stack-data
asyncpg==0.30.0
    # via -r base.in
attrs==25.3.0
    # via
    #   jsonschema
    #   referencing
backcall==0.2.0
    # via ipython
bcrypt==5.0.0
    # via passlib
beautifulsoup4==4.13.5
    # via nbconvert
bleach[css]==6.2.0
    # via nbconvert
build==1.3.0
    # via pip-tools
certifi==2025.8.3
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.3
    # via requests
click==8.3.0
    # via
    #   pip-tools
    #   uvicorn
colorama==0.4.6
    # via
    #   build
    #   click
    #   ipython
    #   tqdm
decorator==5.2.1
    # via ipython
defusedxml==0.7.1
    # via nbconvert
distro==1.9.0
    # via openai
dnspython==2.8.0
    # via email-validator
docopt==0.6.2
    # via pipreqs
email-validator==2.3.0
    # via pydantic
executing==2.2.1
    # via stack-data
fastapi==0.117.1
    # via -r base.in
fastjsonschema==2.21.2
    # via nbformat
greenlet==3.2.4
    # via sqlalchemy
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via openai
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
ipython==8.12.3
    # via pipreqs
jedi==0.19.2
    # via ipython
jinja2==3.1.6
    # via nbconvert
jiter==0.11.0
    # via openai
jsonschema==4.25.1
    # via nbformat
jsonschema-specifications==2025.9.1
    # via jsonschema
jupyter-client==8.6.3
    # via nbclient
jupyter-core==5.8.1
    # via
    #   jupyter-client
    #   nbclient
    #   nbconvert
    #   nbformat
jupyterlab-pygments==0.3.0
    # via nbconvert
mako==1.3.10
    # via alembic
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.2
    # via
    #   jinja2
    #   mako
    #   nbconvert
matplotlib-inline==0.1.7
    # via ipython
mdurl==0.1.2
    # via markdown-it-py
mistune==3.1.4
    # via nbconvert
nbclient==0.10.2
    # via nbconvert
nbconvert==7.16.6
    # via pipreqs
nbformat==5.10.4
    # via
    #   nbclient
    #   nbconvert
openai==1.108.1
    # via -r base.in
packaging==25.0
    # via
    #   build
    #   nbconvert
pandocfilters==1.5.1
    # via nbconvert
parso==0.8.5
    # via jedi
passlib[bcrypt]==1.7.4
    # via -r base.in
pickleshare==0.7.5
    # via ipython
pip-chill==1.0.3
    # via -r dev.in
pip-tools==7.5.0
    # via -r dev.in
pipreqs==0.5.0
    # via -r dev.in
platformdirs==4.4.0
    # via jupyter-core
prompt-toolkit==3.0.52
    # via ipython
psycopg2-binary==2.9.10
    # via -r base.in
pure-eval==0.2.3
    # via stack-data
pydantic[email]==2.11.9
    # via
    #   -r base.in
    #   fastapi
    #   openai
pydantic-core==2.33.2
    # via pydantic
pygments==2.19.2
    # via
    #   ipython
    #   nbconvert
    #   rich
pyjwt==2.10.1
    # via -r base.in
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
python-dateutil==2.9.0.post0
    # via jupyter-client
python-dotenv==1.1.1
    # via -r base.in
pywin32==311
    # via jupyter-core
pyzmq==27.1.0
    # via jupyter-client
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.5
    # via
    #   resend
    #   yarg
resend==2.15.0
    # via -r base.in
rich==14.1.0
    # via -r dev.in
rpds-py==0.27.1
    # via
    #   jsonschema
    #   referencing
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   openai
soupsieve==2.8
    # via beautifulsoup4
sqlalchemy==2.0.43
    # via
    #   -r base.in
    #   alembic
stack-data==0.6.3
    # via ipython
starlette==0.48.0
    # via fastapi
tinycss2==1.4.0
    # via bleach
tornado==6.5.2
    # via jupyter-client
tqdm==4.67.1
    # via openai
traitlets==5.14.3
    # via
    #   ipython
    #   jupyter-client
    #   jupyter-core
    #   matplotlib-inline
    #   nbclient
    #   nbconvert
    #   nbformat
typing-extensions==4.15.0
    # via
    #   alembic
    #   anyio
    #   beautifulsoup4
    #   fastapi
    #   openai
    #   pydantic
    #   pydantic-core
    #   referencing
    #   resend
    #   sqlalchemy
    #   starlette
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
urllib3==2.5.0
    # via requests
uvicorn==0.36.0
    # via -r base.in
wcwidth==0.2.13
    # via prompt-toolkit
webencodings==0.5.1
    # via
    #   bleach
    #   tinycss2
wheel==0.45.1
    # via pip-tools
yarg==0.1.9
    # via pipreqs

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
